Formatting improvements to README.md
I've arranged thumbnail screenshots into a 2x2 table so they fit easier on the page (I wasn't able to do this on the forum). I replaced the Older Versions section with links to where the files are actually stored because I don't think all the details are important enough to have right on the main page (also why I have them in a collapsible "spoiler" section in the forum topic). Finally, I moved the license block to the end, added a link to the forum, and made some minor formatting changes.