Fix problem with some strings showing the wrong color codes.
The root cause here was that strings which different only
in numeric digits were always cached together. However,
digits can also be part of a color code, and strings that
differ in color codes only must be cached separately.

This was first noticed on a PvP server where a player's
name would change team colors between rounds, but the previous
color would be displayed if it was still cached in scrollback
history.