Fix for multithreading issues in Minecraft 1.3.1
When a chat packet arrives in 1.3.1, it's decoded by the TcpReaderThread
which calls getStringWidth(). However, if the passed in string contained
any never before seen glyphs, the GlyphCache would try to render and cache
them into an OpenGL texture. This in turn caused a NullPointerException
in LWJGL.

The solution is to avoid doing any glyph caching at all when called from
another thread. The string layout alone is enough to compute the necessary
width. In addition, the string layout will never be inserted into or
checked against the string cache to avoid concurrent access to the
HashMaps.